requirements.txt
deepagents
langchain-openai
tavily-python
python-dotenv
python-telegram-bot